+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 <!-- Change types:
   ### Added, ### Changed, ### Fixed, ### Removed, ### Deprecated
 -->
+### Fixed
+* User `opts.ignore` extends default opts.
+* Default ignore does not ignore files with .dat in them.
 
 
 ## 3.0.2 - 2016-08-26